+++ src/documents/element-ui.ts
@@ -950,5 +950,89 @@ export default {
 | 参数 | 说明 | 类型 | 可选值 | 默认值 |
 | :--- | :--- | :--- | :--- | :--- |
 | name | 唯一标志符 | string/number | — | — |
-| title | 面板标题 | string | — | — |`
+| title | 面板标题 | string | — | — |`,
+"el-timeline": `[element:https://element.eleme.cn/#/zh-CN/component/timeline](https://element.eleme.cn/#/zh-CN/component/timeline) \n
+| 参数 | 说明 | 类型 | 可选值 | 默认值 |
+| :--- | :--- | :--- | :--- | :--- |
+| reverse | 指定节点排序方向，默认为正序 | boolean | — | false |`,
+"el-timeline-item": `[element:https://element.eleme.cn/#/zh-CN/component/timeline](https://element.eleme.cn/#/zh-CN/component/timeline) \n
+| 参数 | 说明 | 类型 | 可选值 | 默认值 |
+| :--- | :--- | :--- | :--- | :--- |
+| timestamp | 时间戳 | string | — | - |
+| hide-timestamp | 是否隐藏时间戳 | boolean | — | false |
+| placement | 时间戳位置 | string | top / bottom | bottom |
+| type | 节点类型 | string | primary / success / warning / danger / info | - |
+| color | 节点颜色 | string | hsl / hsv / hex / rgb | - |
+| size | 节点尺寸 | string | normal / large | - |
+| icon | 节点图标 | string | — | - |\n
+| name | 说明 |
+| :--- | :--- |
+| - | Timeline-Item 的内容 |
+| dot | 自定义节点 |`,
+"el-divider": `[element：https://element.eleme.cn/#/zh-CN/component/divider](https://element.eleme.cn/#/zh-CN/component/divider) \n
+| 参数 | 说明 | 类型 | 可选值 | 默认值 |
+| :--- | :--- | :--- | :--- | :--- |
+| direction | 设置分割线方向 | string | horizontal / vertical | horizontal |
+| content-position | 设置分割线文案的位置 | string | left / right / center | center |`,
+"el-calendar": `[element：https://element.eleme.cn/#/zh-CN/component/calendar](https://element.eleme.cn/#/zh-CN/component/calendar) \n
+| 参数 | 说明 | 类型 | 可选值 | 默认值 |
+| :--- | :--- | :--- | :--- | :--- |
+| value / v-model | 绑定值 | string | Date/string/number | - |
+| range | 时间范围，包括开始时间与结束时间.开始时间必须是周一，结束时间必须是周日，且时间跨度不能超过两个月。 | Array | - |
+| first-day-of-week | 周起始日 | Number | 1 到 7 | 1 |`,
+"el-image": `[element：https://element.eleme.cn/#/zh-CN/component/image](https://element.eleme.cn/#/zh-CN/component/image) \n
+| 参数 | 说明 | 类型 | 可选值 | 默认值 |
+| :--- | :--- | :--- | :--- | :--- |
+| src | 图片源，同原生 | string | - | - |
+| fit | 确定图片如何适应容器框，同原生 object-fit | string | fill / contain / cover / none / scale-down | - |
+| alt | 原生 alt | string | - | - |
+| referrer-policy | 原生 referrerPolicy | string | - | - |
+| lazy | 是否开启懒加载 | boolean | - | false |
+| scroll-container | 开启懒加载后，监听 scroll 事件的容器 | string / HTMLElement | - | 最近一个 overflow 值为 auto 或 scroll 的父元素 |
+| preview-src-list | 开启图片预览功能 | Array | - | - |
+| z-index | 设置图片预览的 z-index | Number | - | - |\n
+| 事件名称 | 说明 | 回调函数 |
+| :--- | :--- | :--- |
+| load | 图片加载成功触发 | (e: Event) |
+| error | 图片加载失败触发 | (e: Error) |\n
+| slot名称 | 说明 |
+| :--- | :--- |
+| placeholder | 图片未加载的占位内容 |
+| error | 加载失败的内容 |`,
+"el-backtop": `[element：https://element.eleme.cn/#/zh-CN/component/backtop](https://element.eleme.cn/#/zh-CN/component/backtop) \n
+| 参数 | 说明 | 类型 | 可选值 | 默认值 |
+| :--- | :--- | :--- | :--- | :--- |
+| target | 触发滚动的对象 | string | Date/string/number | - |
+| visibility-height | 滚动高度达到此参数值才出现 | number | 200 |
+| right | 控制其显示位置, 距离页面右边距 | number | 40 |
+| bottom | 控制其显示位置, 距离页面底部距离 | number | 40 |\n
+| 事件 | 说明 | 回调参数 |
+| :--- | :--- | :--- |
+| click | 点击按钮触发的事件 | 点击事件 |`,
+"el-drawer": `[element：https://element.eleme.cn/#/zh-CN/component/drawer](https://element.eleme.cn/#/zh-CN/component/drawer) \n
+| 参数 | 说明 | 类型 | 可选值 | 默认值 |
+| :--- | :--- | :--- | :--- | :--- |
+| append-to-body | Drawer 自身是否插入至 body 元素上。嵌套的 Drawer 必须指定该属性并赋值为 true | boolean | - | false |
+| before-close | 关闭前的回调，会暂停 Drawer 的关闭 | function(done)，done 用于关闭 Drawer | - | - |
+| close-on-press-escape | 是否可以通过按下 ESC 关闭 Drawer | boolean | - | true |
+| custom-class	 | Drawer 的自定义类名 | string | - | - |
+| destroy-on-close | 控制是否在关闭 Drawer 之后将子元素全部销毁 | boolean | - | false |
+| modal | 是否需要遮罩层 | boolean | - | true |
+| modal-append-to-body | 遮罩层是否插入至 body 元素上，若为 false，则遮罩层会插入至 Drawer 的父元素上 | boolean | - | true |
+| direction | Drawer 打开的方向 | Direction | rtl / ltr / ttb / tbb | rtl |
+| show-close | 是否显示关闭按钮 | boolean | - | true |
+| size | Drawer 窗体的大小, 当使用 number 类型时, 以像素为单位, 当使用 string 类型时, 请传入 'x%', 否则便会以 number 类型解释 | number / string | - | 30% |
+| title | Drawer 的标题，也可通过具名 slot （见下表）传入 | string | - | - |
+| visible | 是否显示 Drawer，支持 .sync 修饰符 | boolean | - | false |
+| wrapperClosable | 点击遮罩层是否可以关闭 Drawer | boolean | - | true |\n
+| slot名称 | 说明 |
+| :--- | :--- |
+| - | Drawer 的内容 |
+| title | Drawer 标题区的内容 |\n
+| 事件 | 说明 | 回调参数 |
+| :--- | :--- | :--- |
+| open | Drawer 打开的回调 | - |
+| opened | Drawer 打开动画结束时的回调 | - |
+| close | Drawer 关闭的回调 | - |
+| closed | Drawer 关闭动画结束时的回调 | - |`,
 }